Split Competitive Alternatives to its own Component
