+++ src/Modules/Legendaries/LegendaryIcons.js
@@ -1,57 +1,57 @@
 export const legendaryImages = {
-    /* ------------------------------ Druid Legendary Images ------------------------------ */
-    "The Dark Titans Lesson": require("../../Images/Legendaries/TheDarkTitansLesson.jpg"),
-    "Circle of Life and Death": require("../../Images/Legendaries/CircleOfLifeAndDeath.jpg"),
-    "Lycaras Fleeting Glimpse": require("../../Images/Legendaries/LycarasFleetingGlimpse.jpg"),
-    "Memory of the Mother Tree": require("../../Images/Legendaries/MemoryOfTheMotherTree.jpg"),
-    "Oath of the Elder Druid": require("../../Images/Legendaries/OathOfTheElderDruid.jpg"),
-    "Verdant Infusion": require("../../Images/Legendaries/VerdantInfusion.jpg"),
-    "Vision of Unending Growth": require("../../Images/Legendaries/VisionOfUnendingGrowth.jpg"),
-  
-    /* ----------------------------- Paladin Legendary Images ----------------------------- */
-    "Of Dusk and Dawn": require("../../Images/Legendaries/FromDuskTillDawn.jpg"),
-    "Vanguards Momentum": require("../../Images/Legendaries/VanguardsMomentum.jpg"),
-    "The Magistrates Judgment": require("../../Images/Legendaries/TheMagistratesJudgment.jpg"),
-    "Inflorescence of the Sunwell": require("../../Images/Legendaries/InflorescenceOfTheSunwell.jpg"),
-    "Maraads Dying Breath": require("../../Images/Legendaries/MaraadsDyingBreath.jpg"),
-    "Shadowbreaker, Dawn of the Sun": require("../../Images/Legendaries/ShadowbreakerDawnOfTheSun.jpg"),
-    "Shock Barrier": require("../../Images/Legendaries/ShockBarrier.jpg"),
-    "The Mad Paragon": require("../../Images/Legendaries/ability_paladin_conviction.jpg"),
-    "Relentless Inquisitor": require("../../Images/Legendaries/spell_holy_divinepurpose.jpg"),
-  
-    /* ------------------------------ Monk Legendary Images ------------------------------ */
-    "Escape from Reality": require("../../Images/Legendaries/EscapeFromReality.jpg"),
-    "Invoker's Delight": require("../../Images/Legendaries/InvokersDelight.jpg"),
-    "Roll Out": require("../../Images/Legendaries/RollOut.jpg"),
-    "Fatal Touch": require("../../Images/Legendaries/FatalTouch.jpg"),
-    "Ancient Teachings of the Monastery": require("../../Images/Legendaries/AncientTeachingsoftheMonastery.jpg"),
-    "Yu'lon's Whisper": require("../../Images/Legendaries/YulonsWhisper.jpg"),
-    "Clouded Focus": require("../../Images/Legendaries/CloudedFocus.jpg"),
-    "Tear of Morning": require("../../Images/Legendaries/TearofMorning.jpg"),
-    "Vitality Sacrifice": require("../../Images/Legendaries/VitalitySacrifice.jpg"),
-    "Sephuz's Proclamation": require("../../Images/Legendaries/SephuzsProclamation.jpg"),
-    "Echo of Eonar": require("../../Images/Legendaries/EchoofEonar.jpg"),
-  
-    /* ------------------------------ Priest Legendary Images ------------------------------ */
-    "Clarity of Mind": require("../../Images/Legendaries/ClarityofMind.jpg"),
-    "Crystalline Reflection": require("../../Images/Legendaries/CrystallineReflection.jpg"),
-    "Kiss of Death": require("../../Images/Legendaries/KissofDeath.jpg"),
-    "The Penitent One": require("../../Images/Legendaries/ThePenitentOne.jpg"),
-    "Cauterizing Shadows": require("../../Images/Legendaries/CauterizingShadows.jpg"),
-    "Measured Contemplation": require("../../Images/Legendaries/MeasuredContemplation.jpg"),
-    "Twins of the Sun Priestess": require("../../Images/Legendaries/TwinsoftheSunPriestess.jpg"),
-    "Divine Image": require("../../Images/Legendaries/DivineImage.jpg"),
-    "Flash Concentration": require("../../Images/Legendaries/FlashConcentration.jpg"),
-    "Harmonious Apparatus": require("../../Images/Legendaries/HarmoniousApparatus.jpg"),
-    "X'anshi, Return of Archbishop Benedictus": require("../../Images/Legendaries/XanshiReturnofArchbishopBenedictus.jpg"),
-    "Vault of Heavens": require("../../Images/Legendaries/VaultofHeavens.jpg"),
-  
-    /* ----------------------------- Shaman Legendary Images ----------------------------- */
-    "Earthen Harmony": require("../../Images/Legendaries/EarthenHarmony.jpg"),
-    "Jonat's Natural Focus": require("../../Images/Legendaries/JonatsNaturalFocus.jpg"),
-    "Primal Tide Core": require("../../Images/Legendaries/PrimalTideCore.jpg"),
-    "Ancestral Reminder": require("../../Images/Legendaries/AncestralReminder.jpg"),
-    "Chains of Devastation": require("../../Images/Legendaries/ChainsofDevastation.jpg"),
-    "Deeply Rooted Elements": require("../../Images/Legendaries/DeeplyRootedElements.jpg"),
-    "Spiritwalker's Tidal Totem": require("../../Images/Legendaries/SpiritwalkersTidalTotem.jpg"),
-  };
+  /* ------------------------------ Druid Legendary Images ------------------------------ */
+  "The Dark Titans Lesson": require("../../Images/Legendaries/TheDarkTitansLesson.jpg"),
+  "Circle of Life and Death": require("../../Images/Legendaries/CircleOfLifeAndDeath.jpg"),
+  "Lycaras Fleeting Glimpse": require("../../Images/Legendaries/LycarasFleetingGlimpse.jpg"),
+  "Memory of the Mother Tree": require("../../Images/Legendaries/MemoryOfTheMotherTree.jpg"),
+  "Oath of the Elder Druid": require("../../Images/Legendaries/OathOfTheElderDruid.jpg"),
+  "Verdant Infusion": require("../../Images/Legendaries/VerdantInfusion.jpg"),
+  "Vision of Unending Growth": require("../../Images/Legendaries/VisionOfUnendingGrowth.jpg"),
+
+  /* ----------------------------- Paladin Legendary Images ----------------------------- */
+  "Of Dusk and Dawn": require("../../Images/Legendaries/FromDuskTillDawn.jpg"),
+  "Vanguards Momentum": require("../../Images/Legendaries/VanguardsMomentum.jpg"),
+  "The Magistrates Judgment": require("../../Images/Legendaries/TheMagistratesJudgment.jpg"),
+  "Inflorescence of the Sunwell": require("../../Images/Legendaries/InflorescenceOfTheSunwell.jpg"),
+  "Maraads Dying Breath": require("../../Images/Legendaries/MaraadsDyingBreath.jpg"),
+  "Shadowbreaker, Dawn of the Sun": require("../../Images/Legendaries/ShadowbreakerDawnOfTheSun.jpg"),
+  "Shock Barrier": require("../../Images/Legendaries/ShockBarrier.jpg"),
+  "The Mad Paragon": require("../../Images/Legendaries/ability_paladin_conviction.jpg"),
+  "Relentless Inquisitor": require("../../Images/Legendaries/spell_holy_divinepurpose.jpg"),
+
+  /* ------------------------------ Monk Legendary Images ------------------------------ */
+  "Escape from Reality": require("../../Images/Legendaries/EscapeFromReality.jpg"),
+  "Invoker's Delight": require("../../Images/Legendaries/InvokersDelight.jpg"),
+  "Roll Out": require("../../Images/Legendaries/RollOut.jpg"),
+  "Fatal Touch": require("../../Images/Legendaries/FatalTouch.jpg"),
+  "Ancient Teachings of the Monastery": require("../../Images/Legendaries/AncientTeachingsoftheMonastery.jpg"),
+  "Yu'lon's Whisper": require("../../Images/Legendaries/YulonsWhisper.jpg"),
+  "Clouded Focus": require("../../Images/Legendaries/CloudedFocus.jpg"),
+  "Tear of Morning": require("../../Images/Legendaries/TearofMorning.jpg"),
+  "Vitality Sacrifice": require("../../Images/Legendaries/VitalitySacrifice.jpg"),
+  "Sephuz's Proclamation": require("../../Images/Legendaries/SephuzsProclamation.jpg"),
+  "Echo of Eonar": require("../../Images/Legendaries/EchoofEonar.jpg"),
+
+  /* ------------------------------ Priest Legendary Images ------------------------------ */
+  "Clarity of Mind": require("../../Images/Legendaries/ClarityofMind.jpg"),
+  "Crystalline Reflection": require("../../Images/Legendaries/CrystallineReflection.jpg"),
+  "Kiss of Death": require("../../Images/Legendaries/KissofDeath.jpg"),
+  "The Penitent One": require("../../Images/Legendaries/ThePenitentOne.jpg"),
+  "Cauterizing Shadows": require("../../Images/Legendaries/CauterizingShadows.jpg"),
+  "Measured Contemplation": require("../../Images/Legendaries/MeasuredContemplation.jpg"),
+  "Twins of the Sun Priestess": require("../../Images/Legendaries/TwinsoftheSunPriestess.jpg"),
+  "Divine Image": require("../../Images/Legendaries/DivineImage.jpg"),
+  "Flash Concentration": require("../../Images/Legendaries/FlashConcentration.jpg"),
+  "Harmonious Apparatus": require("../../Images/Legendaries/HarmoniousApparatus.jpg"),
+  "X'anshi, Return of Archbishop Benedictus": require("../../Images/Legendaries/XanshiReturnofArchbishopBenedictus.jpg"),
+  "Vault of Heavens": require("../../Images/Legendaries/VaultofHeavens.jpg"),
+
+  /* ----------------------------- Shaman Legendary Images ----------------------------- */
+  "Earthen Harmony": require("../../Images/Legendaries/EarthenHarmony.jpg"),
+  "Jonat's Natural Focus": require("../../Images/Legendaries/JonatsNaturalFocus.jpg"),
+  "Primal Tide Core": require("../../Images/Legendaries/PrimalTideCore.jpg"),
+  "Ancestral Reminder": require("../../Images/Legendaries/AncestralReminder.jpg"),
+  "Chains of Devastation": require("../../Images/Legendaries/ChainsofDevastation.jpg"),
+  "Deeply Rooted Elements": require("../../Images/Legendaries/DeeplyRootedElements.jpg"),
+  "Spiritwalker's Tidal Totem": require("../../Images/Legendaries/SpiritwalkersTidalTotem.jpg"),
+};
