--- src/Modules/TopGear/CloseSetsComponent.js
@@ -1,2 +1,0 @@
-
-// This component shows sets that are similar in strength to the best.
